loopback_test: add porcelain option to command line utility
Marti suggested a user-friendly porcelain option could be added to the
loopback_test utility to make the output more self-documenting and obvious.

This patch adds that porcelain option. When specified output is no longer
produced in CSV format and the CSV file is not touched. Instead output is
directed exclusively to STDOUT with formatting designed to make the output
fields as intutitive as possible to a reader.

The new porcelain option is specified by the "-p" switch on the input
command line.

Example:
loopback_test -t transfer -s 512 -i 32 -p

2015-3-15 22:3:1
 test:			transfer
 path:			/sys/bus/greybus/devices/endo0/
 size:			512
 iterations:		32
 errors:		0
 requests per-sec:	min=133, max=254, average=212.406250, jitter=121
 ap-throughput B/s:	min=140448 max=268224 average=223773  jitter=127776
 ap-latency usec:	min=3929 max=7500 average=4942.781250 jitter=3571
 apbridge-latency usec:	min=3761 max=7315 average=4757.187500 jitter=3554
 gpbridge-latency usec:	min=122 max=123 average=123.187500 jitter=1

2015-3-15 22:3:1
 test:			transfer
 path:			/sys/bus/greybus/devices/endo0:1:2:17:12/
 size:			512
 iterations:		32
 errors:		0
 requests per-sec:	min=133, max=254, average=212.406250, jitter=121
 ap-throughput B/s:	min=140448 max=268224 average=223773  jitter=127776
 ap-latency usec:	min=3929 max=7500 average=4942.781250 jitter=3571
 apbridge-latency usec:	min=3761 max=7315 average=4757.187500 jitter=3554
 gpbridge-latency usec:	min=122 max=123 average=123.187500 jitter=1

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>